# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 디지털 디톡스 앱 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timeslots:
    get:
      operationId: getTimeSlots
      tags: [timeslots]
      summary: Get time slots for a specific day of week
      parameters:
        - in: query
          name: day
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 6
          description: Day of week (0=Sun, 1=Mon, ..., 6=Sat). Defaults to today.
      responses:
        "200":
          description: List of time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
    put:
      operationId: updateTimeSlots
      tags: [timeslots]
      summary: Update time slots (Sunday only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeSlotsRequest"
      responses:
        "200":
          description: Updated time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
        "403":
          description: Not Sunday - settings locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    TimeSlot:
      type: object
      properties:
        index:
          type: integer
        time:
          type: string
        isLocked:
          type: boolean
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
      required: [index, time, isLocked, dayOfWeek]
    UpdateTimeSlotsRequest:
      type: object
      properties:
        day:
          type: integer
          minimum: 0
          maximum: 6
          description: Day of week to update (0=Sun, 1=Mon, ..., 6=Sat)
        slots:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              isLocked:
                type: boolean
            required: [index, isLocked]
      required: [day, slots]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]