# lib/api-zod/src/generated/types/updateTimeSlotsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 디지털 디톡스 앱 API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTimeSlotsRequestSlotsItem } from "./updateTimeSlotsRequestSlotsItem";

export interface UpdateTimeSlotsRequest {
  /**
   * Day of week to update (0=Sun, 1=Mon, ..., 6=Sat)
   * @minimum 0
   * @maximum 6
   */
  day: number;
  slots: UpdateTimeSlotsRequestSlotsItem[];
}